ai/llama2/pyproject.toml
[tool.poetry]
name = "llama2-test"
version = "0.1.0"
description = "My experiments with llama2"
authors = ["Michal Bultrowicz <michal.bultrowicz@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-cpp-python = "^0.2.17"
sentencepiece = "^0.1.99"
gguf = "^0.5.1"


[tool.poetry.group.dev.dependencies]
llama-cpp-python = "^0.2.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
